fixup! settings: normalize link-local and method
For some reason, I thought that as we require the special value "auto"
for mediating with "ipv4.method", that must also be the default value.
That's obviously wrong. The default for the new property can be
"default" just fine, which makes the most sense.

That is, because old uses who upgrade to a new version that supports
ipv4.link-local won't have the global default configured in
NetworkManager.conf. So they will always get the expected default to
"auto", unless the change their configuration.